requirements.txt
asgiref==3.3.1
boto3==1.17.47
certifi==2020.12.5
chardet==4.0.0
dj-static>=0.0.6
dj-stripe==2.4.3
Django==3.1.7
django-base64field==1.0
django-cors-headers==3.2.1
django-storages==1.9.1
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
docutils==0.15.2
fcm-django==0.3.10
geopy==2.1.0
googlemaps==4.4.5
gunicorn==20.1.0
idna==2.10
jmespath==0.10.0
jsonfield==3.1.0
PyJWT==1.7.1
python-dateutil==2.8.1
python-decouple==3.4
psycopg2-binary
pytz==2021.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
static3==0.7.0
stripe==2.56.0
twilio==6.55.0
urllib3==1.25.11
whitenoise==5.2.0
